requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, please pay attention to order them correctly.


falcon==0.1.9
gunicorn>=19.1.0,<=19.2.1
iso8601>=0.1.9
kafka-python>=0.8.1
oslo.config>=1.11.0
oslo.log>=1.0.0
oslo.service>=0.1.0
pastedeploy>=1.3.3
pbr>=0.6,!=0.7,<1.0
python-dateutil>=1.5
six>=1.7.0
stevedore>=0.14
ujson>=1.33
babel
eventlet
pyparsing
